+++ backend/src/modules/client/controllers/dtos/client.dto.ts
@@ -1,0 +1,8 @@
+import { Client } from '@prisma/client';
+
+export class ClientDTO implements Client {
+  id: number;
+  email: string;
+  name: string;
+  phone: string;
+}