+++ backend/src/modules/client/exceptions/client-not-found.exception.ts
@@ -1,0 +1,7 @@
+import { NotFoundException } from '@nestjs/common';
+
+export class ClientNotFoundException extends NotFoundException {
+  constructor() {
+    super('Cliente não encontrado!', ClientNotFoundException.name);
+  }
+}