+++ backend/src/modules/shared/enums/app-controllers.ts
@@ -1,0 +1,4 @@
+export enum AppControllers {
+  app = 'app',
+  client = 'client',
+}