+++ backend/src/modules/shared/enums/app-environment.ts
@@ -1,0 +1,16 @@
+export enum AppEnvironment {
+  nodeEnv = 'NODE_ENV',
+  serviceName = 'npm_package_name',
+  serviceDescription = 'npm_package_description',
+  serviceVersion = 'npm_package_version',
+
+  redisHost = 'REDIS_HOST',
+  redisPort = 'REDIS_PORT',
+  redisPassword = 'REDIS_PASSWORD',
+
+  dbHost = 'DB_HOST',
+  dbUser = 'DB_USERNAME',
+  dbPass = 'DB_PASSWORD',
+  dbName = 'DB_DATABASE',
+  dbLogging = 'DB_LOGGING',
+}