+++ backend/src/modules/shared/enums/swagger-tags.ts
@@ -1,0 +1,4 @@
+export enum SwaggerTags {
+  app = 'Application',
+  client = 'Cliente',
+}