+++ backend/src/modules/shared/services/dispatch-event/interface/http-event-data.interface.ts
@@ -1,0 +1,9 @@
+import { EventType } from './event-type.enum';
+
+export interface HttpEventData {
+  type: EventType.http;
+  authorization?: object;
+  params?: object;
+  headers?: object;
+  body?: object;
+}