+++ CHANGELOG.md
@@ -5,8 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2023-09-20
+
+### Added
+
+- Configurações iniciais do backend com NestJS e Prisma
+- Docker compose com banco de dados e redis
+
 ## [0.0.1] - 2023-09-07
 
 ### Added
+
 - Readme inicial com algumas informações do projeto
 - Diagrama inicial de entidades do sistema